Game: Fix gamepad controls in menu
Apparently, with some platform & joystick combo (most prominent on
Android), pressing the stick and/or D-pad triggers either D-pad keys
or axis movement, instead of hat movement. D-pad keys and axis were
not handled in menu, thus making player unable to navigate the menu.

To fix this, handle D-pad keys and axis in addition to hat movement,
which will work regardless if stick movement is registered as hat,
D-pad or axis.

This also fixes the fact that D-pad and axis was ignored before.